refactor: PathStatus 패키지 위치 변경 및 변경 사항 적용
